renamed the *Object modules
