+++ cffconvert/__init__.py
@@ -1,9 +1,9 @@
-from cffconvert.ApalikeObject import ApalikeObject
-from cffconvert.BibtexObject import BibtexObject
+from cffconvert.apalike import ApalikeObject
+from cffconvert.bibtex import BibtexObject
 from cffconvert.citation import Citation
 from cffconvert.cli import cli
-from cffconvert.CodemetaObject import CodemetaObject
-from cffconvert.EndnoteObject import EndnoteObject
-from cffconvert.RisObject import RisObject
-from cffconvert.SchemaorgObject import SchemaorgObject
-from cffconvert.ZenodoObject import ZenodoObject
+from cffconvert.codemeta import CodemetaObject
+from cffconvert.endnote import EndnoteObject
+from cffconvert.ris import RisObject
+from cffconvert.schemaorg import SchemaorgObject
+from cffconvert.zenodo import ZenodoObject
